Move settings types to settings module
