+++ src/app/useView.ts
@@ -3,8 +3,8 @@ import { get } from "svelte/store";
 import type { DataQueryResult } from "src/custom-view-api";
 import { customViews } from "src/lib/stores/custom-views";
 import type { ViewApi } from "src/lib/view-api";
-import type { ProjectDefinition, ViewDefinition } from "src/types";
 import type { DataRecord } from "src/lib/data";
+import type { ProjectDefinition, ViewDefinition } from "src/settings/settings";
 
 export interface ViewProps {
   view: ViewDefinition;
