+++ src/components/FilterSettings/FilterSettings.svelte
@@ -20,7 +20,7 @@
     type FilterOperator,
     type NumberFilterOperator,
     type StringFilterOperator,
-  } from "src/types";
+  } from "src/settings/settings";
 
   export let filter: FilterDefinition;
   export let fields: DataField[];
