+++ src/settings.ts
@@ -2,7 +2,7 @@ import { App, PluginSettingTab, Setting } from "obsidian";
 import { settings } from "src/lib/stores/settings";
 import { get } from "svelte/store";
 import type ProjectsPlugin from "./main";
-import type { ProjectsPluginPreferences } from "./main";
+import type { ProjectsPluginPreferences } from "./settings/settings";
 
 export class ProjectsSettingTab extends PluginSettingTab {
   constructor(app: App, readonly plugin: ProjectsPlugin) {
@@ -15,6 +15,7 @@ export class ProjectsSettingTab extends PluginSettingTab {
     const save = (prefs: ProjectsPluginPreferences) => {
       settings.updatePreferences(prefs);
     };
+
     const { containerEl } = this;
 
     containerEl.empty();
