+++ src/settings/settings.test.ts
@@ -1,0 +1,7 @@
+import { describe, expect, it } from "@jest/globals";
+
+describe("sortFields", () => {
+  it("sort single field", () => {
+    expect(true).toStrictEqual(true);
+  });
+});
