+++ src/settings/settings.ts
@@ -1,3 +1,5 @@
+import { either } from "fp-ts";
+
 export type ViewType = string;
 
 export interface ViewDefinition {
@@ -156,4 +158,106 @@ export const DEFAULT_PROJECT: UnsavedProjectDefinition = {
     },
   },
   newNotesFolder: "",
+};
+export type ProjectsPluginPreferences = {
+  readonly projectSizeLimit: number;
+  readonly frontmatter: {
+    readonly quoteStrings: "PLAIN" | "QUOTE_DOUBLE";
+  };
+};
+
+export type ProjectsPluginSettingsV1 = {
+  readonly version: 1;
+  readonly lastProjectId?: string | undefined;
+  readonly lastViewId?: string | undefined;
+  readonly projects: ProjectDefinition[];
+  readonly preferences: ProjectsPluginPreferences;
+};
+
+export type ProjectsPluginSettings = ProjectsPluginSettingsV1;
+
+export const DEFAULT_SETTINGS: ProjectsPluginSettings = {
+  version: 1,
+  projects: [],
+  preferences: {
+    projectSizeLimit: 1000,
+    frontmatter: {
+      quoteStrings: "PLAIN",
+    },
+  },
+};
+/**
+ * migrateSettings accepts the value from Plugin.loadData() and returns the most
+ * recent settings. If needed, it applies any necessary migrations.
+ */
+export function migrateSettings(
+  settings: any
+): either.Either<Error, ProjectsPluginSettings> {
+  if (!settings) {
+    return either.right(Object.assign({}, DEFAULT_SETTINGS));
+  }
+
+  if ("version" in settings && typeof settings.version === "number") {
+    // Apply defaults to any saved projects.
+    if ("projects" in settings && Array.isArray(settings.projects)) {
+      return either.tryCatch(() => {
+        return {
+          ...DEFAULT_SETTINGS,
+          ...settings,
+          projects: settings.projects.map(loadProject),
+          preferences: Object.assign({}, DEFAULT_SETTINGS.preferences),
+        };
+      }, either.toError);
+    }
+
+    return either.right({
+      ...DEFAULT_SETTINGS,
+      ...settings,
+    });
+  }
+
+  return either.left(new Error("Missing version"));
+}
+
+// loadProject returns a complete project definition, or throws an exception.
+function loadProject(project: Partial<ProjectDefinition>): ProjectDefinition {
+  const res: UnsavedProjectDefinition = {
+    ...DEFAULT_PROJECT,
+    ...project,
+  };
+
+  if ("name" in res && "id" in res) {
+    const { name, id } = res;
+
+    if (isString(name) && isString(id)) {
+      if ("views" in res && Array.isArray(res.views)) {
+        return { ...res, name, id, views: res.views.map(loadView) };
+      }
+      return { ...res, name, id, views: [] };
+    }
+  }
+
+  throw new Error("Invalid project definition");
+}
+
+// loadProject returns a complete view definition, or throws an exception.
+function loadView(view: Partial<ViewDefinition>): ViewDefinition {
+  const res: UnsavedViewDefinition = {
+    ...DEFAULT_VIEW,
+    ...view,
+  };
+
+  if ("name" in res && "id" in res && "type" in res) {
+    const { name, id, type } = res;
+
+    if (isString(name) && isString(id) && isString(type)) {
+      return { ...res, name, id, type };
+    }
+  }
+
+  throw new Error("Invalid view definition");
+}
+
+const isString = (value: unknown): value is string => {
+  return typeof value === "string";
 };
