+++ src/views/Calendar/CalendarView.svelte
@@ -15,7 +15,7 @@
   import type { ViewApi } from "src/lib/view-api";
   import { CreateNoteModal } from "src/modals/create-note-modal";
   import { EditNoteModal } from "src/modals/edit-note-modal";
-  import type { ProjectDefinition } from "src/types";
+  import type { ProjectDefinition } from "src/settings/settings";
   import {
     fieldToSelectableValue,
     setRecordColorContext,
